requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
pyaes==1.6.1
    # via telethon
pyasn1==0.6.1
    # via rsa
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.3
    # via -r requirements.in
rsa==4.9.1
    # via telethon
telethon==1.42.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
